rename both snippets and fix all the imports
