+++ contents/docs/logs/installation.mdx
@@ -5,7 +5,7 @@ showStepsToc: true
 
 import Tab from "components/Tab"
 import { Steps, Step } from 'components/Docs/Steps'
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/logs-beta.mdx'
 
 <Beta />
