+++ contents/docs/logs/start-here.mdx
@@ -6,7 +6,7 @@ contentMaxWidthClass: max-w-5xl
 
 import { QuestLog, QuestLogItem } from 'components/Docs/QuestLog'
 import { IconCode, IconSearch, IconSettings } from '@posthog/icons'
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/logs-beta.mdx'
 
 <Beta />
