+++ contents/docs/logs/troubleshooting.mdx
@@ -2,7 +2,7 @@
 title: Logs troubleshooting and FAQs
 ---
 
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/logs-beta.mdx'
 
 <Beta />
