+++ contents/docs/workflows/configure-channels.mdx
@@ -4,7 +4,7 @@ showStepsToc: true
 ---
 
 import Tab from "components/Tab"
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/workflows-beta.mdx'
 
 import AddSlackToWorkspace from './_snippets/add-posthog-to-slack-workspace.mdx'
 import AddSlackToChannel from './_snippets/add-posthog-to-slack-channel.mdx'
