+++ contents/docs/workflows/email-drip-campaign.mdx
@@ -2,7 +2,7 @@
 title: "Set up an email drip campaign"
 showStepsToc: true
 ---
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/workflows-beta.mdx'
 
 <Beta />
