+++ contents/docs/workflows/launch-workflow.mdx
@@ -4,7 +4,7 @@ showStepsToc: true
 ---
 
 import Tab from "components/Tab"
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/workflows-beta.mdx'
 
 <Beta />
