+++ contents/docs/workflows/start-here.mdx
@@ -9,7 +9,7 @@ import { IconGraph, IconRewindPlay, IconToggle } from '@posthog/icons'
 import ChannelPlatforms from './_snippets/channel-platforms'
 import OSButton from 'components/OSButton'
 
-import Beta from '../_snippets/beta.mdx'
+import Beta from './_snippets/workflows-beta.mdx'
 
 <Beta />
