+++ contents/docs/workflows/troubleshooting.mdx
@@ -1,7 +1,7 @@
 ---
 title: Workflows troubleshooting and FAQs
 ---
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/workflows-beta.mdx'
 
 <Beta />
