+++ contents/docs/workflows/workflow-builder.mdx
@@ -1,7 +1,7 @@
 ---
 title: Workflow builder
 ---
-import Beta from './_snippets/beta.mdx'
+import Beta from './_snippets/workflows-beta.mdx'
 
 <Beta />
