+++ src/pages/docs/workflows/index.tsx
@@ -8,7 +8,7 @@ import Link from 'components/Link'
 import { ProductScreenshot } from 'components/ProductScreenshot'
 import { Caption } from 'components/Caption'
 import ChannelPlatforms from '../../../../contents/docs/workflows/_snippets/channel-platforms'
-import Beta from '../../../../contents/docs/workflows/_snippets/beta.mdx'
+import Beta from '../../../../contents/docs/workflows/_snippets/workflows-beta.mdx'
 
 export const Content = () => {
     return (
